refactor: move getFamilyMemberProfile to member module
