+++ member/src/main/java/com/oing/controller/MemberController.java
@@ -1,14 +1,34 @@
 package com.oing.controller;
 
-import com.oing.service.MemberService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import com.oing.dto.response.FamilyMemberProfileResponse;
+import com.oing.dto.response.PaginationResponse;
+import com.oing.restapi.MemberApi;
+import org.springframework.stereotype.Controller;
 
-@RestController
-@RequiredArgsConstructor
-public class MemberController {
+import java.util.ArrayList;
+import java.util.List;
 
-    private final MemberService memberService;
+@Controller
+public class MemberController implements MemberApi {
 
+    @Override
+    public PaginationResponse<FamilyMemberProfileResponse> getFamilyMemberProfile(Integer page, Integer size) {
+        String memberIdBase = "01HGW2N7EHJVJ4CJ999RRS2E";
+        String memberNameBase = "디프만";
+
+        List<FamilyMemberProfileResponse> mockResponses = new ArrayList<>();
+        for(int i = 0; i < size; i++) {
+            int currentIndex = i + ((page - 1) * size);
+            String suffix = String.format("%02d", currentIndex);
+            mockResponses.add(
+                    new FamilyMemberProfileResponse(
+                            memberIdBase + suffix,
+                            memberNameBase + suffix,
+                            "https://picsum.photos/200/300?random=" + currentIndex
+                    )
+            );
+        }
+
+        return new PaginationResponse<>(page, 3, size, mockResponses);
+    }
 }
