+++ member/src/main/java/com/oing/restapi/MemberApi.java
@@ -7,18 +7,15 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-@Tag(name = "가족 API", description = "가족 관련 API")
+@Tag(name = "회원 API", description = "회원 관련 API")
 @RestController
 @Valid
-@RequestMapping("/v1/families")
-public interface FamilyApi {
+@RequestMapping("/v1/members")
+public interface MemberApi {
     @Operation(summary = "가족 구성원 프로필 조회", description = "가족 구성원 프로필을 조회합니다.")
-    @GetMapping("/profile")
+    @GetMapping(value = "/profile", params = {"type=FAMILY"})
     PaginationResponse<FamilyMemberProfileResponse> getFamilyMemberProfile(
             @RequestParam(required = false, defaultValue = "1")
             @Parameter(description = "가져올 현재 페이지", example = "1")
